Add files via upload
In this machine learning project, the objective is to predict whether a given sonar signal 
represents a rock or a mine. Logistic regression is a classification algorithm used to predict
the probability of an event occurring. In this project, logistic regression is applied to 
classify sonar signals into two classes: rock and mine.